Enhance equality operator handling and type checking
