+++ src/Runtime.ts
@@ -94,62 +94,3 @@ export class Runtime {
         }
     }
 }
-export function evaluateComparison(operator: string, left: any, right: any): boolean {
-
-    function isValidTimestampString(str: string): boolean {
-        // Accepts ISO 8601 date or datetime strings
-        const parsed = Date.parse(str);
-        if (isNaN(parsed)) return false;
-        // Check if the parsed date's ISO string starts with the input (date or datetime)
-        const d = new Date(parsed);
-        const iso = d.toISOString();
-        // Accepts date-only (YYYY-MM-DD) or full ISO
-        return (
-            str.length === 10 && iso.startsWith(str) ||
-            iso.startsWith(str)
-        );
-    }
-
-    // Convert timestamp strings to Date objects if necessary
-    if (typeof left === 'string' && isValidTimestampString(left)) {
-        left = new Date(left);
-    }
-    if (typeof right === 'string' && isValidTimestampString(right)) {
-        right = new Date(right);
-    }
-
-    // Handle Date objects
-    if (left instanceof Date && right instanceof Date) {
-        const comparison = left.getTime() - right.getTime();
-        switch (operator) {
-            case '<=':
-                return comparison <= 0;
-            case '<':
-                return comparison < 0;
-            case '>=':
-                return comparison >= 0;
-            case '>':
-                return comparison > 0;
-            default:
-                throw new Error(`Unsupported operator: ${operator}`);
-        }
-    }
-
-    // Fallback for numeric or other types
-    if (typeof left === 'number' && typeof right === 'number') {
-        switch (operator) {
-            case '<=':
-                return left <= right;
-            case '<':
-                return left < right;
-            case '>=':
-                return left >= right;
-            case '>':
-                return left > right;
-            default:
-                throw new Error(`Unsupported operator: ${operator}`);
-        }
-    }
-
-    throw new Error(`Relational operator '${operator}' requires operands of the same type (numeric or timestamp), but got '${typeof left}' and '${typeof right}'. Supported types: number, Date`);
-}
